+++ ChangeLog.rst
@@ -1,3 +1,15 @@
+0.35.0 (Unreleased)
+*******************
+
+Security notice
+---------------
+
+- As an extra security measure on top of what the standard Django password reset
+  token generator is already facilitating, allauth now adds the user email
+  address to the hash such that whenever the user's email address changes the
+  token is invalidated.
+
+
 0.34.0 (2017-10-29)
 *******************